requirements.txt
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.1.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
